corrige 15 achados da revisão e mecaniza os gates
Correções de correção:
- apply_trade_allocated: desliga da lista da origem, valida quantidade e I3 antes de mutar,
  rejeita alocação parcial (ADR-0025). O caminho de != para não tinha teste nenhum.
- os laços por conta conferem o dono, em vez de confiar só na lista
- evento corporativo: valida ANTES de consumir a chave de idempotência; switch com default;
  buckets em voo escalam sem gerar sobras (I1 acusava divergência falsa todo dia); subscrição
  valida o teto de direitos
- TradeSettled: outcome validado; BoughtIn de compra devolve a quantidade em vez de perdê-la
- DividendPaid ganha idempotência própria (I6), em conjunto separado
- average_price_half_even devolve zero em posição não positiva em vez de abortar o processo
- Err::ArenaExhausted passa a ser fatal, como o contrato de apply.hpp já dizia
- RV_CHECK: mesma classe nos dois builds (corrupção), em vez de rejeição em release
- GroupHeader de 8 bytes e checagem de alinhamento em view_group
- magic do snapshot de exposição não colide mais com o do segmento do WAL
- dense_index/pair_index: multiplicação alargada antes do transbordo (achado do clang-tidy)

Mecanização:
- CI com matriz de presets, gates do projeto e clang-tidy
- LTO passa a ligar de verdade (PARENT_SCOPE em arquivo include()d ia para lugar nenhum)
- preset tsan passa a selecionar o teste de concorrência, que agora existe
- asan com -fno-sanitize-recover: UBSan reportava e o processo saía 0
- .clang-tidy calibrado, com o motivo de cada check desligado

Testes e simulador:
- tests/base: CRC32C, arena, índices, histograma e o ring SPSC sob dois fios
- tests/domain/test_allocation: seis casos do caminho de != para
- o simulador emite os DEZ eventos e recalcula o net independentemente do motor: zero
  divergência em 105 mil eventos
- o motor-rv-sim publica a soma dos checksums, que não depende do número de partições

+++ src/base/dense_index.hpp
@@ -51,7 +51,10 @@ class DenseIndex {
     uint32_t i = static_cast<uint32_t>(mix64(key)) & mask_;
     for (uint32_t passos = 0; passos <= mask_; ++passos) {
       if (vals_[i] == kEmpty) {
-        if (size_ * 10 >= static_cast<uint64_t>(cap_) * 7) {  // carga > 0,7
+        // `size_ * 10` era calculado em uint32 e SÓ DEPOIS alargado: transborda a partir de
+        // 429 milhões de elementos. Fora de alcance hoje, e exatamente o tipo de limite que
+        // ninguém revisita quando a capacidade cresce. O clang-tidy achou os dois.
+        if (static_cast<uint64_t>(size_) * 10 >= static_cast<uint64_t>(cap_) * 7) {  // carga > 0,7
           inserted = false;
           return kEmpty;
         }
